Add GitHub Actions workflow, Dockerfile, k8s manifests, and related configs
