+++ config/config.json
@@ -3,7 +3,7 @@
     "username": "root",
     "password": "password",
     "database": "ums",
-    "host": "127.0.0.1",
+    "host": "mysql",
     "dialect": "mysql"
   },
   "test": {
