+++ config/database.js
@@ -8,9 +8,11 @@ const sequelize = new Sequelize(
   process.env.DB_USER,
   process.env.DB_PASS, 
   {
-    host: process.env.DB_HOST,
+    host: process.env.DB_HOST ,
     dialect: "mysql",
-    logging: false,
+    port: process.env.DB_PORT ,
+
+    logging: process.env.LOG_LEVEL === "debug" ? console.log : false,
   }
 );
