+++ docker-compose.yml
@@ -1,0 +1,51 @@
+version: "3.8"
+
+services:
+  mysql:
+    image: mysql:8
+    container_name: ums_mysql
+    restart: unless-stopped
+    environment:
+      MYSQL_ROOT_PASSWORD: $DB_PASS
+      MYSQL_DATABASE:  $DB_NAME
+    ports:
+      - "3306:3306"
+    volumes:
+      - db_data:/var/lib/mysql
+    networks:
+      - backend
+    healthcheck:
+      test: ["CMD", "mysqladmin", "ping", "-h", "localhost", "-u", "root", "-p$$MYSQL_ROOT_PASSWORD"]
+      timeout: 30s
+      retries: 10
+      start_period: 40s
+
+  app:
+    build: .
+    container_name: ums_app
+    restart: on-failure
+    ports:
+      - "8080:8080"
+    env_file:
+      - .env
+    depends_on:
+      mysql:
+        condition: service_healthy
+    networks:
+      - backend
+    command: >
+      sh -c "
+        echo 'Waiting for database to be ready...' &&
+        until nc -z mysql 3306; do echo 'Waiting for MySQL...'; sleep 2; done &&
+        echo 'Database is ready!' &&
+        npx sequelize-cli db:migrate &&
+        npx sequelize-cli db:seed:all &&
+        npm start
+      "
+
+volumes:
+  db_data:
+
+networks:
+  backend:
+    driver: bridge
