+++ k8s/app-deployment.yaml
@@ -1,0 +1,32 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: ums-app
+spec:
+  replicas: 2
+  selector:
+    matchLabels:
+      app: ums-app
+  template:
+    metadata:
+      labels:
+        app: ums-app
+    spec:
+      containers:
+        - name: ums-app
+          image: ums_app:latest 
+          imagePullPolicy: Never
+          ports:
+            - containerPort: 8080
+          env:
+            - name: DB_HOST
+              value: mysql
+            - name: DB_USER
+              value: root
+            - name: DB_PASSWORD
+              valueFrom:
+                secretKeyRef:
+                  name: mysql-secret
+                  key: mysql-root-password
+            - name: DB_NAME
+              value: ums_db
