+++ k8s/app-service.yaml
@@ -1,0 +1,13 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: ums-app-service
+spec:
+  type: NodePort
+  selector:
+    app: ums-app      
+  ports:
+  - port: 8080
+    targetPort: 8080
+    nodePort: 30081
+
